code_interpolation/requirements.txt
cffi==1.11.2
cycler==0.10.0
decorator==4.1.2
fpdf==1.7.2
functools32==3.2.3.post2
imageio==2.2.0
matplotlib==2.0.2
moviepy==0.2.3.5
networkx==2.0
numpy==1.13.1
opencv-python==4.0.0.21
pillow==4.1.1
protobuf==3.4.0
pycparser==2.18
pyparsing==2.2.0
python-dateutil==2.6.0
pytz==2017.2
pywavelets==0.5.2
pyyaml==3.12
requests==2.19.1
scikit-image==0.13.1
scipy==0.18.1
six==1.10.0
subprocess32==3.2.7
tensorboardX==0.8
torch==0.3.1
torchvision==0.2.0
seaborn==0.9.0
pandas==0.20.3
prettytable==0.7.2
